ofc unsigned cannot be <0

+++ src/lcd/nextion.cpp
@@ -412,7 +412,7 @@ void lcdFetchPage(eepromValues_t &settings, NextionPage page, int targetProfile)
 
 uint8_t lcdGetSelectedProfile(void) {
   uint8_t pId;
-  uint8_t attempts = 2;
+  int attempts = 2;
   do {
     if (attempts-- <= 0) {
       lcdShowPopup((String("getProfile rekt: ") + pId).c_str());
